+++ src/options/cli-options.ts
@@ -6,6 +6,7 @@ interface ParsedArgs extends minimist.ParsedArgs {
   help: boolean
   stdin: boolean
   version: boolean
+  ['stdin-filename']?: string
   globals?: string | string[]
   plugins?: string | string[]
   envs?: string | string[]
@@ -17,6 +18,7 @@ interface ParsedArgs extends minimist.ParsedArgs {
 export interface CLIOptions {
   fix: boolean
   useStdIn: boolean
+  stdInFilename?: string
   files?: string[]
   project?: string
   globals?: string[]
@@ -37,7 +39,15 @@ export function getCLIOptions (): CLIOptions {
       project: 'p'
     },
     boolean: ['fix', 'help', 'stdin', 'version'],
-    string: ['globals', 'plugins', 'parser', 'envs', 'project', 'report']
+    string: [
+      'globals',
+      'plugins',
+      'parser',
+      'envs',
+      'project',
+      'report',
+      'stdin-filename'
+    ]
   }) as ParsedArgs
 
   // Unix convention: Command line argument `-` is a shorthand for `--stdin`
@@ -58,17 +68,18 @@ Usage:
     files/folders that begin with '.' like .git/) are automatically ignored.
     Paths in a project's root .gitignore file are also automatically ignored.
 Flags:
-        --fix       Automatically fix problems
-    -p, --project   Specify ts-config location (default: ./tsconfig.eslint.json or ./tsconfig.json)
-        --version   Show current version
-    -h, --help      Show usage information
+        --fix              Automatically fix problems
+    -p, --project          Specify ts-config location (default: ./tsconfig.eslint.json or ./tsconfig.json)
+        --version          Show current version
+    -h, --help             Show usage information
 Flags (advanced):
-        --stdin     Read file text from stdin
-        --globals   Declare global variable
-        --plugins   Use custom eslint plugin
-        --envs      Use custom eslint environment
-        --parser    Use custom ts/js parser (default: @typescript-eslint/parser)
-        --report    Use a built-in eslint reporter or custom eslint reporter (default: standard)
+        --stdin            Read file text from stdin (requires using --stdin-filename)
+        --stdin-filename   The filename and path of the contents read by stdin
+        --globals          Declare global variable
+        --plugins          Use custom eslint plugin
+        --envs             Use custom eslint environment
+        --parser           Use custom ts/js parser (default: @typescript-eslint/parser)
+        --report           Use a built-in eslint reporter or custom eslint reporter (default: standard)
     `)
     return process.exit(0)
   }
@@ -82,9 +93,10 @@ Flags (advanced):
   // Get the files/globs to lint
   const files = argv._.length !== 0 ? argv._ : undefined
 
-  return {
+  const options = {
     fix: argv.fix,
     useStdIn: argv.stdin,
+    stdInFilename: argv['stdin-filename'],
     files,
     project: argv.project,
     globals: exports._convertToArray(argv.globals),
@@ -93,6 +105,15 @@ Flags (advanced):
     parser: argv.parser,
     report: argv.report
   }
+
+  if (options.useStdIn && options.stdInFilename == null) {
+    console.error(
+      'Must provide the `--stdin-filename` flag when using the `--stdin` flag.'
+    )
+    process.exit(1)
+  }
+
+  return options
 }
 
 export function _convertToArray<T> (val?: T | T[]): T[] | undefined {